Take DocFx build back out of pipeline
Since DocFx still uses the full .NET FX, we can't really run this on
Ubuntu agents. (In theory it works with Mono, but it's not clear whether
you can get that on hosted build agents.)

DocFx 3 is supposed to be shipping some time this year, which should
render this moot, so if we can avoid dealing with it, we should.

In any case, we need to sort out our cross-repo doc build, and we might
just end up relying on that to validate docs. (It'd be better to get an
earlier warning. We'd prefer to block merges if the doc build is
failing, but we'll pass on that for now.)