Make CancellationToken arguments optional
There are two cases where I've not done this:

* internal types
* public implementations of ITokenSource

The reason for the latter is that the normal use for these types is to pass them to some library that requires an ITokenSource. Nothing should be calling the relevant methods directly on those classes - they will be going via the interface. (As it happens, ITokenSource itself (not our interface) doesn't define a default value for cancellation tokens, but that's not our problem to fix.)