Tidying up various issue missed in an earlier PR (#215)
* Remove unnecessary full namespace qualifications
* Change complex expression-bodied method back to full method
* Make CancellationToken arguments optional
* Fix typos
* Fix unfinished paragraph in configuration.md
* Add detail about what "dynamic" configuration means

There are two exceptions to the optionalization of cancellation tokens:
* internal types
* public implementations of ITokenSource

The reason for the latter is that the normal use for these types is to pass them to some library that requires an ITokenSource. Nothing should be calling the relevant methods directly on those classes - they will be going via the interface. (As it happens, ITokenSource itself (not our interface) doesn't define a default value for cancellation tokens, but that's not our problem to fix.)