Add ClientIdentityConfiguration-based cache invalidation
Previously, to be able to invalidate cached data for credentials, you had to hang onto the particular IAccessTokenSource, IAzureTokenCredentialSource etc that you were returned, and invoke its GetReplacementForFailed... method. This turned out to be problematic because Corvus.Storage wasn't actually holding onto those instances, and to be able to do so introduced a lot of complexity.

There wasn't really any need to require this because the cache operates at a higher scope anyway, and only really needs to see what's in the ClientIdentityConfiguration to know what to invalidate.

So this change enables client code to take advantage of that, asking us to invalidate and cached data associated with a particular ClientIdentityConfiguration without needing to tie it to a particular object.