Enable Azure CLI, VS, and VS Code credentials (#274)
We used to have to use AzureIdentityDefaultAzureCredential to take advantage of either Azure CLI or VS credentials, which was unnecessarily slow, because it would always first try to acquire a managed identity, only trying alternatives after timing out.

Also, we had no support for VS code before. (Although the VS Code Azure Account extension appears to have changed how it works in ways Azure.Identity has not yet caught up with - you need to be on an old version of the extension if you want this to work right now. This isn't something we can control - the exact same issue exists on Python because the basic problem is in the VS Code extension.)